* merged commit from https://github.com/dkimitsa/robovm/commits/launcher-rework
* fixed: idea hang when running debug session and process failed before debugger got disconnected from robovm-debugger code (e.g. while deploying to simulator)
Root case: once process terminated idea was waiting for stdout/error stream to be closed (as process is dead it should happen). But there was a ProcessProxy used around the process and it uses stdout/err streams connected to real process via file FIFOs as result it was not a process stream but file stream to FIFO which was never closed once process was dead. This caused a hang.
The Fix:
- launcher processes unified: single abstract process and thin implementation
- FIFOs removed as streams can be easily routed using piped streams;
- launch code re-arranged: there were two copies in AppCompier and Abstract target. Only one left.
- ProcessProxy removed as not required anymore